requirements.txt
fastapi==0.103.0
uvicorn==0.21.1
pillow==9.4.0
numpy==1.24.4
paddlepaddle==2.6.2
paddleocr==2.9.1
pandas==2.0.2
setuptools==41.2.0
python-multipart==0.0.20